+++ _layouts/parrot/TopBar/TopBar.scss
@@ -74,28 +74,3 @@
   }
 }
 
-.sideMenu {
-  position: fixed;
-  top: 0;
-  left: 0;
-
-  background-color: transparent;
-  transition: background-color 300ms ease-out;
-
-  &.visible {
-    @include color(background-color, background-overlay, 0.6);
-  }
-  &.closed {
-    left: -100%;
-  }
-
-  height: 100%;
-  width: 100%;
-
-  display: flex;
-
-  @include break(700, 900) {
-    display: none;
-  }
-}
-
